Fix some casting

+++ Source/WubVocoder/PitchShifter.cpp
@@ -10,13 +10,13 @@ void WubVocoderPitchShifter::init (int order)
 {
     stft.init (order);
     istft.init (order);
-    windowlength = pow (2.0f, order);
-    fftdata.assign (windowlength, 0.0f);
-    fftmag.assign (windowlength, 0.0f);
-    fftphase.assign (windowlength, 0.0f);
-    oldphase.assign (windowlength, 0.0f);
-    modphase.assign (windowlength, 0.0f);
-    deltaphase.assign (windowlength, 0.0f);
+    windowlength = (int) powf (2.0f, order);
+    fftdata.assign ((size_t) windowlength, 0.0f);
+    fftmag.assign ((size_t) windowlength, 0.0f);
+    fftphase.assign ((size_t) windowlength, 0.0f);
+    oldphase.assign ((size_t) windowlength, 0.0f);
+    modphase.assign ((size_t) windowlength, 0.0f);
+    deltaphase.assign ((size_t) windowlength, 0.0f);
     firsttime = true;
 }
 
@@ -28,7 +28,7 @@ void WubVocoderPitchShifter::step (float* samples, int numsamples, float hoprati
     // Do strange and wonderful things in the frequency domain
     float pi = 3.14159f;
 
-    for (int n = 0; n < windowlength; n++)
+    for (size_t n = 0; n < (size_t) windowlength; n++)
     {
         // Separate magnitude and phase, saving previous phase
         oldphase[n] = fftphase[n];
@@ -54,7 +54,7 @@ void WubVocoderPitchShifter::step (float* samples, int numsamples, float hoprati
     // Calculate ISTFT using modified block size
     int synthlength = int (round (hopratio * float (numsamples)));
     std::vector<float> synthsamples;
-    synthsamples.assign (synthlength, 0.0f);
+    synthsamples.assign ((size_t) synthlength, 0.0f);
     istft.step (fftdata.data(), synthsamples.data(), synthlength);
 
     // Resample to original block size
