+++ Source/WubVocoder/stft.cpp
@@ -4,12 +4,14 @@
 void STFT::init (int order)
 {
     windowlen = int (pow (2.0f, order));
-    hannwindow.assign (windowlen, 0.0f);
-    sampleframe.assign (windowlen, 0.0f);
-    fftinput.assign (windowlen, 0.0f);
+    hannwindow.assign ((size_t) windowlen, 0.0f);
+    sampleframe.assign ((size_t) windowlen, 0.0f);
+    fftinput.assign ((size_t) windowlen, 0.0f);
+
     float pi = 3.14159f;
     for (int n = 0; n < windowlen; n++)
-        hannwindow[n] = sqrt (0.5f * (1.0f - cosf (2.0f * pi * float (n) / float (windowlen))));
+        hannwindow[(size_t) n] = sqrt (0.5f * (1.0f - cosf (2.0f * pi * float (n) / float (windowlen))));
+
     fft = std::make_unique<dsp::FFT> (order);
 }
 
@@ -17,12 +19,15 @@ void STFT::step (float* samples, int numsamples, Cfloat* fftoutput)
 {
     // Shift samples into frame
     for (int n = numsamples; n < windowlen; n++)
-        sampleframe[n - numsamples] = sampleframe[n];
+        sampleframe[(size_t) (n - numsamples)] = sampleframe[(size_t) n];
+
     for (int n = 0; n < numsamples; n++)
-        sampleframe[windowlen - numsamples + n] = samples[n];
+        sampleframe[(size_t) (windowlen - numsamples + n)] = samples[n];
+
     // Apply window
-    for (int n = 0; n < windowlen; n++)
+    for (size_t n = 0; n < (size_t) windowlen; n++)
         fftinput[n] = sampleframe[n] * hannwindow[n];
+
     // Compute FFT
     fft->perform (fftinput.data(), fftoutput, false);
 }
